requirements.txt
faster-whisper
sounddevice
webrtcvad
keyboard
pynput
pyperclip
numpy
customtkinter
pystray
pillow
ollama
pygetwindow
pyinstaller
